requirements.txt
itchat
echarts-python
shodan